ReportDeadBodyをPlayerControlに統合
